refactor: remove supporter plus thank you
